Create a SelectDeviceFeature components which can be re-used
