+++ front/src/routes/scene/edit-scene/triggers/DeviceFeatureState.jsx
@@ -1,57 +1,23 @@
 import { Component } from 'preact';
 import { connect } from 'unistore/preact';
 import { Text, Localizer } from 'preact-i18n';
-import Select from 'react-select';
 import cx from 'classnames';
+
+import SelectDeviceFeature from '../../../../components/device/SelectDeviceFeature';
 
 @connect('httpClient', {})
 class TurnOnLight extends Component {
-  getOptions = async () => {
-    try {
-      const rooms = await this.props.httpClient.get('/api/v1/room?expand=devices');
-      const deviceOptions = [];
-      rooms.forEach(room => {
-        const roomDeviceFeatures = [];
-        room.devices.forEach(device => {
-          device.features.forEach(feature => {
-            roomDeviceFeatures.push({
-              value: feature.selector,
-              label: feature.name,
-              type: feature.type,
-              unit: feature.unit
-            });
-          });
-        });
-        if (roomDeviceFeatures.length > 0) {
-          roomDeviceFeatures.sort((a, b) => {
-            if (a.label < b.label) {
-              return -1;
-            } else if (a.label > b.label) {
-              return 1;
-            }
-            return 0;
-          });
-          deviceOptions.push({
-            label: room.name,
-            options: roomDeviceFeatures
-          });
-        }
-      });
-      await this.setState({ deviceOptions });
-      this.refreshSelectedOptions(this.props);
-      return deviceOptions;
-    } catch (e) {
-      console.log(e);
-    }
-  };
-  handleChange = selectedOption => {
-    if (selectedOption && selectedOption.value) {
-      this.props.updateTriggerProperty(this.props.index, 'device_feature', selectedOption.value);
-      this.props.updateTriggerProperty(this.props.index, 'value', null);
+  onDeviceFeatureChange = deviceFeature => {
+    this.setState({ selectedDeviceFeature: deviceFeature });
+    if (deviceFeature) {
+      this.props.updateTriggerProperty(this.props.index, 'device_feature', deviceFeature.selector);
+      if (deviceFeature.selector !== this.props.trigger.device_feature) {
+        this.props.updateTriggerProperty(this.props.index, 'value', null);
+      }
     } else {
       this.props.updateTriggerProperty(this.props.index, 'device_feature', null);
     }
-    if (selectedOption && selectedOption.type === 'binary') {
+    if (deviceFeature && deviceFeature.type === 'binary') {
       this.props.updateTriggerProperty(this.props.index, 'operator', '=');
     }
   };
@@ -67,24 +33,6 @@ class TurnOnLight extends Component {
   };
   handleValueChangeBinary = newValue => () => {
     this.props.updateTriggerProperty(this.props.index, 'value', newValue);
-  };
-  refreshSelectedOptions = nextProps => {
-    let selectedOption = '';
-    if (nextProps.trigger.device_feature && this.state.deviceOptions) {
-      let deviceOption;
-      let i = 0;
-      while (i < this.state.deviceOptions.length && deviceOption === undefined) {
-        deviceOption = this.state.deviceOptions[i].options.find(
-          option => option.value === nextProps.trigger.device_feature
-        );
-        i++;
-      }
-
-      if (deviceOption) {
-        selectedOption = deviceOption;
-      }
-    }
-    this.setState({ selectedOption });
   };
   getBinaryOperator = () => (
     <div class="col-2">
@@ -125,32 +73,22 @@ class TurnOnLight extends Component {
       </div>
     </div>
   );
-  constructor(props) {
-    super(props);
-    this.state = {
-      deviceOptions: null,
-      selectedOption: ''
-    };
-  }
-  async componentDidMount() {
-    this.getOptions();
-  }
-  componentWillReceiveProps(nextProps) {
-    this.refreshSelectedOptions(nextProps);
-  }
 
-  render(props, { selectedOption, deviceOptions }) {
+  render(props, { selectedDeviceFeature }) {
     return (
       <div>
         <div class="row">
           <div class="col-6">
             <div class="form-group">
-              <Select defaultValue={''} value={selectedOption} onChange={this.handleChange} options={deviceOptions} />
+              <SelectDeviceFeature
+                value={this.props.trigger.device_feature}
+                onDeviceFeatureChange={this.onDeviceFeatureChange}
+              />
             </div>
           </div>
-          {selectedOption && selectedOption.type === 'binary' && this.getBinaryOperator()}
-          {selectedOption && selectedOption.type === 'binary' && this.getBinaryButtons()}
-          {selectedOption && selectedOption.type !== 'binary' && (
+          {selectedDeviceFeature && selectedDeviceFeature.type === 'binary' && this.getBinaryOperator()}
+          {selectedDeviceFeature && selectedDeviceFeature.type === 'binary' && this.getBinaryButtons()}
+          {selectedDeviceFeature && selectedDeviceFeature.type !== 'binary' && (
             <div class="col-3">
               <div class="form-group">
                 <select class="form-control" onChange={this.handleOperatorChange} value={props.trigger.operator}>
@@ -177,7 +115,7 @@ class TurnOnLight extends Component {
               </div>
             </div>
           )}
-          {selectedOption && selectedOption.type !== 'binary' && (
+          {selectedDeviceFeature && selectedDeviceFeature.type !== 'binary' && (
             <div class="col-3">
               <div class="form-group">
                 <div class="input-group">
@@ -190,10 +128,10 @@ class TurnOnLight extends Component {
                       onChange={this.handleValueChange}
                     />
                   </Localizer>
-                  {selectedOption && selectedOption.unit && (
+                  {selectedDeviceFeature.unit && (
                     <span class="input-group-append" id="basic-addon2">
                       <span class="input-group-text">
-                        <Text id={`deviceFeatureUnitShort.${selectedOption.unit}`} />
+                        <Text id={`deviceFeatureUnitShort.${selectedDeviceFeature.unit}`} />
                       </span>
                     </span>
                   )}
